better cats transform feature.
