+++ include/EquipScene.hpp
@@ -7,7 +7,7 @@
 
 #include "FrameScene.hpp"
 #include "GameButton.hpp"
-#include "UnitCard.hpp"
+#include "CatList.hpp"
 
 class App;
