+++ include/UpgradeScene.hpp
@@ -8,7 +8,7 @@
 #include "pch.hpp"
 #include "FrameScene.hpp"
 #include "GameButton.hpp"
-#include "UnitCard.hpp"
+#include "CatList.hpp"
 
 class UpgradeScene : public FrameScene{
 public:
